fix(payment-gateway): keep monetary values as BigDecimal in promotion flow
Remove double conversions from purchase promotion requests and results by using BigDecimal end-to-end, preventing floating-point precision loss in currency calculations.

Made-with: Cursor

+++ backend/commerce-platform/services/payment-gateway/src/main/java/bbmovie/commerce/payment_gateway/client/dto/PromotionApplyRequest.java
@@ -1,8 +1,10 @@
 package bbmovie.commerce.payment_gateway.client.dto;
 
+import java.math.BigDecimal;
+
 public record PromotionApplyRequest(
         String code,
         String userId,
-        Double cartValue
+        BigDecimal cartValue
 ) {
 }
